Refactor - First Pass!
New classes for server, settings, logger & strategy
Moved a lot of market functions to public/private

Strategy class defines a base strat with rainbow like
defaults, that can be overriden. There are simple examples for
each of the main EE strats, these are optimised to be stable and
try to keep playing whatever happens.

Still very much work in progress :)